pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = "Projeto de lista de tarefas com FastAPI"
authors = ["Henrique Sebastião <contato@henriquesebastiao.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = {extras = ["standard"], version = "^0.112.1"}
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
python-multipart = "^0.0.9"
pyjwt = "^2.9.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.7"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
isort = "^5.13.2"
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations', '__init__.py']

[tool.ruff.lint]
preview = true
select = ['F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.isort]
profile = "black"
line_length = 79
extend_skip = ['migrations']

[tool.blue]
extend-exclude = '(migrations/)'

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'

[tool.taskipy.tasks]
run = 'fastapi dev fast_zero/app.py'
pre_test = 'task lint'
test = 'pytest --cov=fast_zero -vv'
post_test = 'coverage html'
cov = 'firefox-developer-edition htmlcov/index.html'
lint = 'ruff check . && blue --check . --diff && isort --check diff .'
format = 'blue . && isort .'